Added Console API
This is a self made API that is still in testing. What this aims to do is
act as an easy accessor to the console for ease of output. It treats the
console as a grid of characters to be printed. This will allow for better
printing of images and other ASCII art to the terminal.